导入方法抽取到 BaseTreeTab，清除 contacts/calendar 重复代码
- BaseTreeTab 新增 _import_add_item / _import_refresh_list 接口
- base_tab 实现 _import_selected / _import_file / _import_url / _import_clipboard / show_text_import / show_import_preview / _open_import_preview
- contacts_tab / calendar_tab 删除重复的导入方法，改为继承 BaseTreeTab 共享实现
